Learning Management System
This project aims to replace human labor using a desktop application. The current assessment of students is being done manually, which is a difficult for instructors. The system is designed to perform
CRUD operations on a database schema consisting of 10 tables: Assessment, StudentResult, AssessmentComponent, Student, RubricLevel, StudentAttendance, Lookup, Rubric, ClassAttendance, and
Clo. The Assessment table contains information about the assessments. The StudentResult table stores
the results of the students’ assessments. The Student table stores the students’ personal information.
The RubricLevel table defines the rubric levels. The StudentAttendance table tracks the attendance of
the students. The ClassAttendance table tracks the attendance of the classes. The system also generates
PDF reports that provide useful information to the instructors. Overall, this project aims to demonstrate the use and benefits of databases in real-world problems, such as managing large amounts of data
efficiently and securely.